moved all styling to styles/*.scss (except About.js because I'm afraid of divine retribution)
