+++ weddingbells/src/styles/faq.scss
@@ -1,0 +1,3 @@
+.faq-wrapper {
+    height: 500px;
+}
