+++ weddingbells/src/styles/header.scss
@@ -1,0 +1,40 @@
+.header-wrapper {
+	// border
+	// border: 1px red solid;
+
+	// sizing & placement
+	position: fixed;
+	top: 0px;
+	width: 100%;
+	height: 80px;
+
+	// background
+	background-color: #c4c4c4;
+
+	// display
+	display: flex;
+	flex-direction: row;
+	flex-wrap: wrap;
+	justify-content: space-between;
+	align-items: center;
+
+	// text
+	font-size: 23px;
+}
+
+.header-wrapper .header-logo {
+	// size & placement
+	margin-left: 75px;
+}
+
+.header-wrapper .header-links {
+	// size & placement
+	margin-right: 75px;
+	width: 30%;
+
+	// display
+	display: flex;
+	flex-direction: row;
+	flex-wrap: wrap;
+	justify-content: space-between;
+}
